backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
transformers==4.47.1
torch==2.5.1
torchaudio==2.5.1
webrtcvad==2.0.10
google-generativeai==0.8.4
elevenlabs==1.50.5
authlib==1.4.0
httpx==0.28.1
python-multipart==0.0.18
pydub==0.25.1
numpy==1.26.4
scipy==1.14.1
